Fix Libinjection can not detect 1.e or 1.E SQLI issue
https://github.com/libinjection/libinjection/issues/59

https://gosecure.ai/blog/2021/10/19/a-scientific-notation-bug-in-mysql-left-aws-waf-clients-vulnerable-to-sql-injection/
From this Bolg, we could see 1.e or 1.E is a risk SQLI, The SQL parser will ignore it during parsing.

Like "1.e(1)" => (1), 1 1.e/1 => 1/1 etc.
So, if a payload such as "1' or 1.e(1)" bypass SQLI detection, which is really risk,
we should detect such SQLI injection in case WAF bypass.
The easiest sloution maybe just ignore it instead of parsing it as TYPE_BAREWORD , the SQLI parser will continue to move to get the value after 1.e

Why?
First, in some context, the 1.e definitly should be parsed as TYPE_BAREWORD, but it no harmful if we ignore it even though it should be parsed as TYPE_BAREWORD
Please think about the below cases:

E.g
input_data = token1 token2 1.2e token4 token5 token6 ....

Pasing xx.xxE as TYPE_BAREWORD

case 1: if input_data detects a SQLI, then it maybe a "false positive" detection if we parse 1.2e as TYPE_BAREWORD, since it could not be a SQLI after remove 1.2e. [we need fix]
case 2: if input_data not detects a SQLI, then it maybe a "false negative" detection if we parse 1.2e as TYPE_BAREWORD, since it could be a SQLI after remove 1.2e. [we need fix]

Ignoring xx.xxE

case 1: if input_data detects a SQLI, which is a "true positive" detection not false. Because after SQLI pass to SQL parser, xxx.xxE will be ignored, if we parsed it as TYPE_BAREWORD, then we may get a "false negative" detection [Fix the above case 2]
case 2: if input_data not detects a SQLI, which is a "true negative" detection. Why ? Let's assume it's a SQLI, after pass it to SQL parser, 1.2e will be ignored, and then execute the rest of SQL, but the rest of is not a valid SQL or SQLI. But if we parse 1.2e as TYPE_BAREWORD(n), then we may get a "false positive" detection, since xx.xxE should be ignored. [Fix the above case 1]

Examples:
1' or 1.e(1)
abc or 2 1.e%2
1' or 2 1.E/1
1' or 2 1.e%1
1' or 2 1.E^1
1' or 2 1.e|1
1' or 2 1.e*1
1' or 2 1.e&1
1' or 2 1.e/id
a select char 10.2E(id 2.e)
1 or 1.e(1)